refactor(console-ui): phase 5c — earn feature module (F7)
Split the 780-line earn page (calc.ts already held the model). New module:
- useEarningsCalculator.ts: all hardware/model selection state + the useMemo
  derivations + handlers (the orchestration that was tangled into the page).
- SetupProviderCTA / HardwareSelector / ModelSelector / EarningsResults /
  PillButton: dumb presentation over the hook result.
- page.tsx: thin orchestrator (780 -> ~70 lines).

Clears the earn/page max-lines guardrail flag. Invariant held: build green,
eslint 0 errors / 114 warnings, vitest 358 pass / 4 pre-existing fail
(earn-page.test still green).
